pyproject.toml
[tool.poetry]
name = "robo dan"
version = "0.1.0"
description = ""
authors = ["Void-ux <daniel.grant2017@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = {extras = ["speed"], version = "^2.0.1"}
Pillow = "^9.2.0"
toml = "^0.10.2"
aiohttp = {extras = ["speed"], version = "^3.8.1"}
asyncpg = "^0.26.0"
colorama = "^0.4.5"
psutil = "^5.9.2"
Unidecode = "^1.3.4"
python-dateutil = "^2.8.2"
parsedatetime = "^2.6"
requests = "^2.28.1"
jishaku = "^2.5.1"
yt-dlp = "^2022.9.1"
aiob2 = "^0.4.4"
mystbin-py = "^5.0.2"
orjson = "^3.8.0"
uvloop = { version = "^0.16.0", markers = "sys_platform == 'linux'" }
paramiko = "^2.11.0"
scp = "^0.14.4"
rich = "^12.5.1"
humanize = "^4.3.0"
wavelink = "^1.3.2"
maturin = "^0.13.3"

[tool.poetry.dev-dependencies]

[tool.pyright]
venvPath = "C:\\Users\\MS1\\AppData\\Local\\pypoetry\\Cache\\virtualenvs"
venv = "overseer-aiZiDOpq-py3.10"
typeCheckingVersion = "off"
pythonVersion = "3.10"
reportUnusedImport = "error"
reportUnnecessaryTypeIgnoreComment = "error"
reportIncompatibleMethodOverride = "error"

[build-system]
requires = ["maturin>=0.13,<0.14"]
build-backend = "maturin"